pyproject.toml
[project]
name = "pdf-to-audiobook"
version = "0.1.0"
description = "Generate audiobook audio file from pdf"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.14.3",
    "ffmpy>=1.0.0",
    "markdown>=3.10",
    "marker-pdf[full]>=1.10.1",
    "tts>=0.22.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.8",
]
